Implemented GUI in WPF
Changed class to Sorter.cs and made it compile to a .dll
ImageSort is the WPF UI.

This is extremely inefficient... The garbage collector does
not delete the stored information for a long time... This
program (Release build) was using 1.7 GB of RAM at one point,
since I ran it 3 times in a row. It didn't garbage collect
until after that, at which point it was only using 11 KB.

It probably has to do with my code, but I'm going to write this
in C++ with QT (with the intent to make it cross-platform), and
I will be able to manually manage memory.